Actors walks on top of map tiles
Refactored movement from Player* to new superclass ActorSpriteNode.

Actors will now stand on top of map tiles (with the assumption that the top surface is the isometric bounding box for the tile). No more walking inside walls and such ghostly nonsense.

Actors has a limited step height that they can reach when moving, and can't move between tiles with greater height difference (as measured in base tile halfwidth). This should probably be replaced with an ActionPoint cost function of step height in the future.

Signed-off-by: Calle Englund <calle@discord.bofh.se>