Randomized map generation
Refactored lots of tile related stuff, so the tile templates can hold more information than just a texture image name, e.g. the stepHeight for a tile, so we can have tiles that are not perfectly flat, but still be walkable and have actors positioned correctly.

Loaded some http://opengameart.org/ tile artwork to get something to work with for randomized map generation.

Signed-off-by: Calle Englund <calle@discord.bofh.se>